+++ choco-solver/src/main/java/org/chocosolver/solver/search/limits/TimeCounter.java
@@ -57,7 +57,7 @@ public class TimeCounter extends ACounter implements IMonitorOpenNode {
 
     @Override
     public void init() {
-        solver.getMeasures().updateTimeCount();
+        solver.getMeasures().updateTime();
         long time = (long) (solver.getMeasures().getTimeCount()*1000f);
         offset = System.currentTimeMillis() - time;
     }