+++ choco-solver/src/main/java/org/chocosolver/solver/trace/LogStatEveryXXms.java
@@ -55,8 +55,7 @@ public class LogStatEveryXXms implements IMonitorInitPropagation, IMonitorClose 
                     sleep(duration);
                     //noinspection InfiniteLoopStatement
                     do {
-                        solver.getMeasures().updateTimeCount();
-                        solver.getMeasures().updatePropagationCount();
+                        solver.getMeasures().updateTime();
                         Chatterbox.out.println(String.format(">> %s", solver.getMeasures().toOneShortLineString()));
                         sleep(duration);
                     } while (alive);